option for p.file to mds.single
now we can send a .p file from MrBayes, loaded as a table, to the
mds.single function.

This is only used for making the plot, in which case we can plot the
LnL and the mcmc.sample number on the same plot.

Need to add in a bunch of exceptions for when the .p file isn't there.
Right now there will be issues.