remove check for unrooted trees rom tree.dist.matrix.liam
because it didn't seem to work on multiphylo lists. For reasons I
didn't figure out yet, a list of unrooted trees returns a list of
TRUE's when you do this:

sapply(unclass(trees),is.rooted))

This is not the intended behaviour. And in any case, we should move all
of this to a function that reads in trees and does these checks once
only.