add N option to tree.ess.R, to specify number of thinnings
this used to be hardcoded at 20, but testing suggests that increasing
it a bit can improve accuracy a lot.

There are smarter methods, but I'm not that inclined to bother with
them. E.g. you could find the first thinning that crosses the line,
then search between that and the next for the true crossing point, and
iterate.