Create parse.clades
Created parse.clades.R, containing the parse.clades function.  This
function takes a vector of clades given as tip numbers (e.g., the
cladetable attribute of many of our analyses) and a multiphylo object
and returns a vector of clades with those numbers converted to names.
Names are taken from the multiphylo object.