Various fixes to load.trees
Added functionality to make it take non-nexus tree format.  Also added
bits to make it unroot trees.  Because testing for rootedness is
time-consuming on large tree files, I've set it up so that it just
checks the first tree in the file and assumes that are all unrooted or
rooted if that one is.