Fixed plot.tree.ess
This was barfing erros when called from the tree.ess function, but not
when it was called manually.  Turns out it the tree.distances attribute
was being called (and others) before it was added to the object.  I
modified it so that it works from tree.ess now.  To call it from the
console you need to directly feed it the tree.distances table.