Documentation changes, new sample data
Added some new sample data (now two chains with p tables).  Updated the
documentation so that all sample code is executable, which is necessary
for compiling the package.  Did a bunch of trial building to find
errors, and fixed most of those.  There’s still a few warnings (see
00check.log in the rwty.Rcheck directory), but I’m probably not going
to mess with those today.