fix bug in treespace plots
some combination of the way that we were naming the trees, and the hack
of adding a small number to zero distances was creating a nasty bug.
Specifically, we could feed in a list of identical trees, and get a
treespace plot that looked like a nice ball.

Anyway, but is now fixed in tree.dist.matrix