Made merge.ptables work for chains of different length
The merge.ptables function works for chains of different length now, so
the only restriction is that they need to have the same number of gens
per tree.  However, there are functions downstream that don’t like
chains of different length so this is only the first step in the
process of making RWTY okay with that.