Posterior probability plots working with analyze.simple
Got everything hooked up so that posterior plots are now working.  Got
rid of analyze.multi, going to move all of that functionality into
plot.multichain.

As it stands, analyze.single is still doing all of the work, but we’ll
probably change the name of that.  It’s also returning the data, but
that’s not being stuffed into the final object built from
analyze.simple.  I’m doing this so that we/users can still call
analyze.single manually if we/they want that data.