Beast log files
Setting up code to auto-locate beast log files.  The test data I have
from Beast has some weird naming scheme because it’s a multilocus
analysis, so I’m not sure how generally workable this is going to be.
Anyway, it just checks to see if the extension is .t, and if so tries
for a .p log file, or if the extension is .trees it looks for .log.