Bean plots for ESS
Added bean plots for ESS.  In order to do so, I changed the way
topological.ess returns results - it now spits back a complete table of
results instead of summary statistics.  Even if we decide to go back to
the dot and whisker plots, I think we should do things this way - with
the new version, end users can directly access the topo ESS stuff much
easier in case they need to do that for some reason.