Rendered plots to images
Turned plots into images for RMD, storing them now in the images
directory.  GitHub apparently doesn’t make plots on the fly for RMD
documents, or I’m doing it wrong.  Possible that I need to use plot()
for that, but RMD was barfing on the compare plots using that command.