added function to get tree distances from first tree after burnin
this is used for plotting the topology trace. It does exactly what the
pseudo ESS function does, except that we default to using the first
tree after burnin as the focal tree, because it seems like a sensible
default for a plot. This is because if we use the first tree, we can
recognise burnin as a climb to a plateau in the same way that we are
used to seeing for e.g. lnL plots.