added treedist option to topology analysis functions
we now give the option to calculate tree distances with path distance
or robinson-foulds.

But, in the plots (which most people will use) we do not give the
option. This is because I see no reason why people should prefer RF
over PD. The reason they’re included here is for research into
comparing them, not diagnosis of MCMC chains.