update focal tree for makeplot.topology.trace
Hopefully the final change. We now use the final tree of the burnin of
the first chain. This usually gets rid of the potentially misleading
artefacts caused by using a tree that appears in one of the chains as
the focal tree.

if burnin = 0, we just use the first tree of the first chain, since ALL
trees will be in the analysis somewhere.