begin to fix compare.n
this had a number of issues. The code was a bit of a mess. The plot
labels didn’t work, and the plot using GGally took a really long time.
I decided to go for a home-baked solution, which currently lacks a bit
of the finesse of the GGally plot, but has the advantage of being
almost instantaneous to plot.

Still working on getting some text into the upper-right hand plots.