Fixed issue when only one topology in 95% CI of topo space
Thanks to Arun Prasanna for bringing this to our attention.  There was
a bug in makeplot.pairs that was causing RWTY to barf when the chain
wasn’t exploring enough topologies to have any variation in the 95% CI
of topology space.  Basically it was trying to color a histogram with
categories of zero width.

It’s still not producing density plots for correlations between
topology and other parameters under these circumstances, but those
would basically be a vertical line anyway.  Might be worth revisiting,
but for now it’s at least functional.