Fixing various bugs
Fixed an error about parameter correlations showing up when no param
table was provided.

Also had a screwy and unnecessary … call to prop.part, got rid of that.

Fixed an issue with parse.clades that was calling the tip labels from
trees incorrectly and causing an error.  This one is interesting not
because it barfed an error on some data sets, but because there were
some where it worked perfectly and I have no idea why or how.