tree distance matrix
This is the start of a major refactor, where for each chain we start by calculating a full tree distance matrix.

The point is to then never calculate distances again, except for the tree topology plot which for now is an exception (because for that to work we'd need to calculate ALL by ALL tree distances across chains and that seems like a waste of time).