fix issues with autocorrelation and ess
specifically:

fix a bug with burnin = 0

remove user ability to specify maximum sampling interval - that was useful when calculations were slow, but now we can just have it be the maximum value for which we can get 20 samples.

And if we can get fewer than 10 sampling intervals with at least 20 samples, then there are at most 30 trees so we just report ESS's that are < the number of trees.